+++ src/lib.rs
@@ -2,4 +2,5 @@
 
 #![forbid(unsafe_code)]
 
+pub mod adapter;
 pub mod vehicle_state;